requirements.txt
pandas==1.5.3
numpy==1.24.4
xarray==2023.1.0
geopandas==0.13.2
fiona==1.9.6
netCDF4==1.6.5
h5netcdf==1.1.0
numexpr==2.8.6
PyYaml==6.0.2
